requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
environ==1.0
sqlparse==0.5.1
